bq_view/requirements.txt
google-cloud-bigquery==3.12.0
jsonargparse[signatures]==4.27.0
